修改了gpsbug但是环形缓冲区有问题

+++ Core/GPS/GPS.h
@@ -2,14 +2,14 @@
 #define GPS_h
 #include "RingBuff.h"
 #include <stdio.h>
-
+#include "usart.h"
 //此位gps的数据结构体 
 typedef struct {    
     uint8_t lockType;         // GPS定位锁定类型
     uint8_t lockFlags;        // GPS锁定标志位
     uint8_t satelliteCount;   // 用于导航解算的卫星数量
-    int32_t longitude;        // 经度（单位：degree）
-    int32_t latitude;         // 纬度（单位：degree）
+    double longitude;        // 经度（单位：degree）
+    double latitude;         // 纬度（单位：degree）
     int32_t velocityNorth;    // 北方向速度（单位：mm/s）
     int32_t velocityEast;     // 东方向速度（单位：mm/s）
     int32_t velocityDown;     // 向地速度（单位：mm/s）
